Stop fullscreening link label for resources/settings
The link label was resizing itself as large as the panel, which mean that mouse over color occured as soon as the mouse moved anywhere in the property page. Centered the link label and changed it's size to match the text it contains.
Also stopped showing focus cues by default, unless you actually used the keyboard, it was distracting.