Add telemetry logging for unexpected exceptions in Microsoft.VisualStudio.Editors
Logging Infrastructure: We are using the VS Telemetry APIs to log actionable exceptions, this is the recommended approach across DevDiv for Dev15.

Added API: Added a utility API "ReportWithoutCrash" that given an exception, logs telemetry for the exception with context (name of the throwing component and additional context message) and returns whether or not the exception is recoverable. API has optional flags to trigger Debug.Fail or treat given exception as always recoverable (catch all filter).

Code changes in the project to consume this API: API is always invoked in the exception filter for catch clauses. Cases where this API invocation was added:

1. Debug.Fail within the catch clause: We surely care about telemetry for these cases.
2. Existing invocations of Utils.IsUnrecoverableException in catch filters where telemetry might help fix/cleanup code: Used my best judgement on whether telemetry data is helpful, erring on conservative side of more data when I was unsure.
3. [WIP] Silently swallowing System.Exception without rethrowing. It is recommended to catch specific exception kinds for scenarios where we know an operation might throw. However, there are tons of such occurences across the project, and it is not clear if we going to clean up all of them. I have changed few with invocations to ReportWithoutCrash so we know the exact exception being thrown, but fixing over the entire project would need someone familiar with the code to do an audit.

Related issue: https://github.com/dotnet/roslyn-project-system/issues/41