Cross targeting: Projects build all TFMs from VS
Ensure that for cross targeting projects we clear out the "TargetFramework" global property during explict builds in VS so that we build for all configurations. We do so by implementing a build-only properties provider that checks if the project configuration specifies "TargetFramework" as a dimension or not, and clear out the TargetFramework property if the dimension exists.

Fixes #504