Stop doubling scaling MaximumSize/MinimumSize
The Debug property page was manually scaling the MaximumSize and MinimumSize properties - yet these are automatically scaled by the WinForms layout engine. This was resulting in the Debug property page in being forced into either too large or too size than normal.

I suspect this is the root cause of #427, but been unable to actually get VS in a state where it will down scale the property pages enough to repro. It looks like to do this would require a combination of a smaller Icon font size (which VS uses for its font) plus a smaller DPI than 96 (something Win10 no longer lets you do). This requires more investigation and plan on finding a machine in this state.