Fix "System cannot be referenced" in Error List
After #7a8718d, a warning was introduced in the Error List around "System not being able to referenced". This occured due to the way the legacy NuGet build task adds/removes framework assemblies for project.json. It does it during design-time builds which confuses csproj, and results it spitting out a warning. This was occuring to the toolset project only due to the unneeded framework asemblies coming from the analyzers references (https://github.com/dotnet/roslyn-analyzers/issues/1092).

Adding a new dependency type as a workaround of this issue.