Bump up the version of our assemblies from 2.0.0.XXX to 2.0.1.XXX
Post RC.2 we fixed our versioning to use the correct revision field, but this causes RC.2 version to be treated as higher then our latest version. This causes the VS upgrade from RC.2 to RC.3 to not upgrade our Project System and Editor assemblies. To workaround this, we are bumping up Build portion of the version from 0 to 1.