Get Property after context for GeneralBrowseObject
When CPS creates DynamicProperty, the backing object, that provides property value to the property pages, it obtains InContext value and Final value for each property. InContext value is the value of the property uptill the evaluation of the proj file. For OutputPath property, it is empty since OutputPath is defined in the .targets, which gets evaluated after the project file. The final value for OutputPath is the expected value.

During the creation of the DynamicProperty, the backing DataSource can either say pick the InContext value or the Final Value. The default is InContext(represented as BeforeContext in code). This change makes the data source to use the AfterContext value/final value, which is what the user is interested in.