Add binding redirect for Immutable for tests
Tests were failing inside VS due to Immutable mismatch - CPS expects version 1.2, but due to MSBuild dependency we deploy 1.2.1.

Make note this wasn't causing us to fail command-line runs because they run inside the xUnit appdomain (due to xunit.AppDomain == denied) and looks like (according to fuslog) it's failing back to a LoadFrom bind to the exact binding in the directory next to the test when the Load bind fails.