Fix CoreCompile is never up-to-date
Fixes: #825

There are certain situations where MSBuild will always call Csc/Vbc regardless of whether it's inputs/outputs are up-to-date. One these occasions is when it's building the project and UseHostCompilerIfAvailable is set to true (see _ComputeNonExistentFileProperty in Microsoft.Common.targets).

For CPS projects (and out-of-proc legacy project system builds), never use the host compiler - we always shell out builds to the process VBCSCompiler, or csc/vbc if that fails.