Prevent deadlock when the project file has conditioned TargetFrameworks property
We currently do not support conditioned TargetFrameworks property, due to the the restriction of CPS/MSBuild to do a project evaluation at unconfigured project level. However, we should not deadlock VS when such a project file is provided. This fix ensures that we avoid infinite workspace project context computation when evaluated TargetFrameworks doesn't match the known project configurations.

Fixes #1431