Ensure that the LangaugeServiceHost is initialized when attempting to fetch the contained langauge factory for a venus file.
We are hitting a race condition in test automation runs for ASP.Net core projects, where if a cshtml file in a web project is opened before the project has finished loading, the langauge service host is not initialized and hence GetContainedLanguageFactoryForFile might fail. The fix is to ensure that we explicitly await langauge service host initialization at the beginning of GetContainedLanguageFactoryForFile.