Project.EditProjectFile should be always available
Project.EditProjectFile was only enabled when the root node was selected in Solution Explorer, and Solution Explorer had focus. Instead, it should be enabled regardless of which node is selected, this makes make it available when editor has focus or from the Project when menu one or many project items are selected. This is how other project-level commands work such as Add Reference.

This doesn't cause it to appear on the context menu of an item - by default it doesn't show up unless the user adds it to that context menu.