Fix for 1965 Attempting to show properties of certain references displays an error about access being denied
When supplied "file" parameter for properties context, CPS tried to use that as suffix to append to project folder when evaluating project item for properties. That some whow ended up in creating real folders on disk (i will check with CPS team why is that) and when user did not have elevated VS we had access violation exceptions since CPS tried to create files under current folder , which sometimes could be system32 etc.