Listen to Namespace import changes via subscription instead of getting it from from MSBuild
Instead of getting the information from the Live MSBuild object to answer questions about VB Namespace Imports subscribe to the changes to the Namespace imports and maintain a list, which can be used to answer questions about the import. MSBuild Object will be used only to update the MSBuild.

This change also handles some corner case and throw exception accordingly