Show all analyzers in Solution Explorer
Currently when expanding the "Analyzers" node in Solution Explorer you will only see analyzers that have been added directly to the project--analyzers from NuGet packages will not be shown. To get to those you would need to dig through the "NuGet" node and then the individual packages to find what you want.

A major reason we show analyzers in the Solution Explorer is to allow the user to expand them, see the diagnostics each one supports, and adjust the severity. Forcing the user to dig through the NuGet packages would make this effectively impossible as the user would need to know ahead of time which packages include which analyzers, or else they have to spend a lot of time clicking around until they find what they want.

To simplify this experience, we should just show all the analyzers that are going to be passed to the compiler in a flat list. We accomplish this by adding a new target, `CollectAnalyzersDesignTime`, and updating the `AnalyzerReference` and `ResolvedAnalyzerReference` rules to depend on that.