Stop resetting workspace context for single TFM changes
Partially fixes #2148

Basically, we're resetting the workspace context here - but we're not resetting the project subscription, so we don't get sent the original data again so the new context is missing the stuff that didn't change.

There's really zero point to resetting the workspace context on TFM change, it was done to workaround some issues, instead I'm going to tackle those underlying issues.