Add CSharp and VB Project Configuration Properties
These are the Project Configuration Properties which backs the Automation Objects obtained through DTE.

Currently these properties are exported as the automation object. In the future, this implementation would be exported as an object, which will be com aggregated with the ProjectConfig Object in CPS. A new Component will be added in the Managed Project System, which will import ProjectConfig and export that object as the Automation Object. With the new approach, the ProjectConfig will be the object thats powers both automation and IVsCfgBrowseObject. This behavior will be in sync with the Native Project system.