Remove explict handling of empty project changes
#0f1c164b added explicit handling of empty project changes with zero context as to why it was added, other than it happens during design-time build failure.

I've tried all that I can to repro this situation - I cannot; breaking builds due to evaluation breaks, failing all targets, failing just a single target, switching configurations and a few other cases never cause CPS to send on an empty set of changes. Code inspection on the CPS side leads me to believe this an impossible situation.

However, if we were to run into this situation - I want to explicitly handle the diff similar to what we did in #4abcdc8 to avoid the workspace getting out of sync, and hence added an assertion to make us aware of the issue if we run into it again in the future.

Replace the check for an empty set of project changes with an assert. I looked back through the history and the bugs that were fixed when this change was introduced and cannot repro a situation where CPS sends us an empty set of project changes. I've tried lots of situations to reply

If this actually does occur:
- We'll want to specially handle this situation better - because the next set of diffs are going to be broken and we'll be out of sync.