Propery extract and check the assembly name
When checking if an assembly-qualified name refers to the "netstandard" assembly we're currently only checking that the substring "netstandard" occurs _somewhere_. This is incorrect, as we'll then redirect unrelated types that happen to have that text somewhere in their name.

This initial check should only serve to bail us out early and avoid allocations in the common case that the type has nothing to do with netstandard and shouldn't be redirected. We need a further check where we actually parse out an `AssemblyName` instance and explicitly check the `Name` property against "netstandard".