Remove synchronous call in DependencySubscriptionsHost
Fixes #2379

Replace the synchronous call to read the project file, and replaced it with reading from snapshot data.

There were two issues:

1) We were blocking a ThreadPool thread with a synchronous call, which was leading to ThreadPool exchaustion.

2) We were reading from the live project instead of the snapshot of the project - these can be out of sync.

We read the project file to figure if a given item is imported or lives in the project file ("implicit"). This affects a few things, including it's icon (which we change to a lock) and the ability to delete it.

Note, there are a few entrypoints into this method, some of which have project snapshot data (from the evaluation.design-time build) and some of which don't (when we get an event from a referenced project indicating that it changed).

The simplest approach without pushing async through all events (which requires a public API change of an API used by Web Tooling), was just to push a null set of project items indicating what we "no data" and avoid flipping/setting the "implicit" setting when another project changes it's results.