Simplify FindByPathInternal
In CPS project tree terms, the dependency node looks like this (with only ever 3 nodes deep)

Dependencies
    Analyzers
        Analyzer
    Projects
        Reference

We were attempting to walk into the child of Analyzer and Reference - even though that code never gets hit.

Replaced this walk with just a call to GetSelfAndDescendentsBreadthFirst - it's simplier and ends up resulting in the exact same number of allocations (5) as we're not walking each child's children twice.