Use NewTree overloads available prior to 15.5
Commit 334dcf6a updated us to consume a 15.5 version of the CPS
binaries. This includes changes to support specifying a sort order on
tree nodes, needed by F#.

When we moved over our calls in the Dependencies tree code
to the `NewTree` method started binding to new overloads that can take
a value indicating a sort order, and return an instance of the new
`IProjectTree2` or `IProjectTreeItem2` interfaces. However, these new
overloads aren't available at runtime on VS builds prior to 15.5,
leading to "method not found" exceptions when trying to run locally
built project-system bits on top of a 15.4 Preview VS.

Since there are no publicly available 15.5 builds and the Dependencies
tree doesn't need to specify its own sort order, this commit updates
the calls so we will use the overloads available prior to 15.5.