Clean up the logic to save the imports in the VB references page.
Fixes internal bug 509520 - an automatic crash dump for an IndexOutOfRangeException.

I *suspect*, though can't tell from the crash dump I've been looking at that the issue here
is that the VSLangProj "Imports" collection has a different number of items than what is
passed to the compiler and retrieved from the compilation.

To fix, I cleaned up the logic, and stopped relying on the data from the compiler, instead
just collecting the items from the VSLangProj "Imports".