Add AvoidPersistingProjectGuidStorageProvider
An implement of IProjectGuidStorageProvider that replaces our project rule for <ProjectGuid /> property. This avoids us needing to evaluate to figuring out the project's GUID - and also avoids us from needed to take a read-lock when we are subsequentially told to set the project GUID.

The rule for ProjectGuid will be removed in a later commit.