Avoid loading additional configurations for multi-targeting projects
To determine the configurations when loading a project, we need to read the project for the dimensions (the <Configurations>, <Platforms> and <TargetFrameworks> properties). This leads to a chicken and egg problem; to figure out all configurations for a given project you need to load at least one configuration to evaluate. This results us loading an empty configuration (called "NoVariablesDefault") just to grab the actual real configurations, see: #2711.

To avoid unnecessarily loading this unneeded config, we "guess" what the defaults will be based on reading the Configuration, Platform & TargetFramework dimensions via the MSBuild construction model without evaluating.