Avoid picking implicit framework defines
Fixes: #2720

The SDK automatically appends to DefineConstants with implicit constants after the project's definition of DefineConstants, when we read the value at the end of project file ("AfterContext") we think that user's value is set to include those when we re-read it.

Instead move to reading the value *at the point* it was defined in the project file, which causes us to ignore framework implicit defines.