Avoid picking up implicit output path
Fixes: #2177

The SDK automatically appends to OutputPath with implicit framework and runtime values, when we attempt to read the value at the end of project file ("AfterContext") we think that user's value is set to include those and pick these up in the dialog which sets the property.

Instead move to reading the value at the point it was defined in the project file, which causes us to ignore any values that the SDK appends.