Do not fail IVsQueryDebuggableProjectCfg.QueryDebugTargets when called on library
When called on a library, IVsQueryDebuggableProjectCfg.QueryDebugTargets was failing preventing Test Explorer from pulling the current debug settings.

QueryDebugTargetsAsync is called in two situations:

- When launching
- When someone queries the available debug targets

Similar to legacy project system, only fail when called during launch so that we can continue to show an error on F5'ing a library.

Fixes: #3647.

Note, we now fall back the target path similar if RunCommand is not specified. This brings the behavior in line with legacy when called on a library.