Avoid blocking thread on releasing write lock
Dispose synchronously blocks the thread until the write lock is released. Under the covers the project service yields on this path, which under thread pool exhaustion situations can cause Dispose to be blocked for a long time, exacerbating the problem. Await on ReleaseAsync so that the current thread isn't blocked can be released back to the pool.