Introduce a DependencyIconSet type
Currently every `Dependency` and `DependencyModel` stores four `ImageMoniker`s specifying the set of icons associated with the dependency. As each `ImageMoniker` is a struct containing a `GUID` and an `int` , each `Dependency` and `DependencyModel` ends up with at least 80 bytes of icon information.

While effectively each `Dependency` instance is unique there aren't that many distinct icon sets--just one or two per type of dependency. We can thus save a fair bit of memory by breaking the icons out into a new type, `DependencyIconSet`, replacing all 80 bytes of `ImageMonikers`  with references to those sets, and reusing instances of `DependencyIconSet` where possible.

This commit introduces the `DependencyIconSet` type and updates the `Dependency` type to make use of it. It also reuses `DependencyIconSet` instances where it is trivial to do so.