Reuse icon sets
When creating a `Dependency` from another `Dependency` or a
`DependencyModel` we can reuse the `DependencyIconSet` contained in the
original instance, rather than copying the `ImageMoniker`s into a new
icon set.

This covers the common cases where then `IDependencyModel` implementation
passed to the `Dependency` constructor is one of ours. If the
implementation comes from another component (for example, the components
that provide `IDependencyModel`s for NPM and Bower packages) we'll still
need to copy the `ImageMoniker`s.